organize addForwarder tests
