+++ test/index.spec.ts
@@ -1,30 +1,11 @@
 import * as lambda from "@aws-cdk/aws-lambda";
-import { CfnSubscriptionFilter } from "@aws-cdk/aws-logs";
-
 import * as cdk from "@aws-cdk/core";
 import "@aws-cdk/assert/jest";
-import { addForwarder } from "../src/forwarder";
 import { Datadog, logForwardingEnvVar, enableDDTracingEnvVar, injectLogContextEnvVar } from "../src/index";
 import { JS_HANDLER_WITH_LAYERS, DD_HANDLER_ENV_VAR, PYTHON_HANDLER } from "../src/redirect";
-const SUBSCRIPTION_FILTER_PREFIX = "DatadogSubscriptionFilter";
+import { findDatadogSubscriptionFilters } from "./test-utils";
 
-function findDatadogSubscriptionFilters(stack: cdk.Stack) {
-  return stack.node
-    .findAll()
-    .filter((construct) => construct.node.id.startsWith(SUBSCRIPTION_FILTER_PREFIX))
-    .map((construct) => {
-      const cfnSubscriptionFilter = construct.node.children.find(
-        (child) => child instanceof CfnSubscriptionFilter,
-      ) as CfnSubscriptionFilter;
-
-      return {
-        id: construct.node.id,
-        destinationArn: cfnSubscriptionFilter.destinationArn,
-      };
-    });
-}
-
-describe("addForwarder", () => {
+describe("addLambdaFunctions", () => {
   it("Subscribes the same forwarder to two different lambda functions via separate addLambdaFunctions function calls", () => {
     const app = new cdk.App();
     const stack = new cdk.Stack(app, "stack", {
@@ -54,35 +35,11 @@ describe("addForwarder", () => {
     datadogCdk.addLambdaFunctions([nodeLambda]);
     datadogCdk.addLambdaFunctions([pythonLambda]);
 
-    const subscriptionFilters = findDatadogSubscriptionFilters(stack);
-    expect(nodeLambda.logGroup.node.tryFindChild(subscriptionFilters[0].id)).not.toBeUndefined();
-    expect(pythonLambda.logGroup.node.tryFindChild(subscriptionFilters[1].id)).not.toBeUndefined();
-    expect(subscriptionFilters[0].destinationArn).toEqual(subscriptionFilters[1].destinationArn);
-  });
-
-  it("Subscribes the same forwarder to two different lambda functions via one addForwarder function call", () => {
-    const app = new cdk.App();
-    const stack = new cdk.Stack(app, "stack", {
-      env: {
-        region: "sa-east-1",
-      },
-    });
-    const nodeLambda = new lambda.Function(stack, "NodeHandler", {
-      runtime: lambda.Runtime.NODEJS_10_X,
-      code: lambda.Code.fromAsset("test"),
-      handler: "hello.handler",
-    });
-    const pythonLambda = new lambda.Function(stack, "PythonHandler", {
-      runtime: lambda.Runtime.PYTHON_3_7,
-      code: lambda.Code.fromAsset("test"),
-      handler: "hello.handler",
-    });
-    addForwarder(stack, [nodeLambda, pythonLambda], "forwarder-arn");
-
-    const subscriptionFilters = findDatadogSubscriptionFilters(stack);
-    expect(nodeLambda.logGroup.node.tryFindChild(subscriptionFilters[0].id)).not.toBeUndefined();
-    expect(pythonLambda.logGroup.node.tryFindChild(subscriptionFilters[1].id)).not.toBeUndefined();
-    expect(subscriptionFilters[0].destinationArn).toEqual(subscriptionFilters[1].destinationArn);
+    const [nodeLambdaSubscriptionFilter] = findDatadogSubscriptionFilters(nodeLambda);
+    const [pythonLambdaSubscriptionFilter] = findDatadogSubscriptionFilters(pythonLambda);
+    expect(nodeLambdaSubscriptionFilter).toBeDefined();
+    expect(pythonLambdaSubscriptionFilter).toBeDefined();
+    expect(nodeLambdaSubscriptionFilter.destinationArn).toEqual(pythonLambdaSubscriptionFilter.destinationArn);
   });
 
   it("Throws an error when a customer redundantly calls the addLambdaFunctions function on the same lambda function(s) and forwarder", () => {
@@ -115,81 +72,8 @@ describe("addForwarder", () => {
     }
     expect(throwsError).toBe(true);
   });
-
-  it("Subscribes two different forwarders to two different lambda functions via separate addForwarder function calls", () => {
-    const app = new cdk.App();
-    const stack = new cdk.Stack(app, "stack", {
-      env: {
-        region: "sa-east-1",
-      },
-    });
-    const nodeLambda = new lambda.Function(stack, "NodeHandler", {
-      runtime: lambda.Runtime.NODEJS_10_X,
-      code: lambda.Code.fromAsset("test"),
-      handler: "hello.handler",
-    });
-    const pythonLambda = new lambda.Function(stack, "PythonHandler", {
-      runtime: lambda.Runtime.PYTHON_3_7,
-      code: lambda.Code.fromAsset("test"),
-      handler: "hello.handler",
-    });
-    const datadogCdk = new Datadog(stack, "Datadog", {
-      nodeLayerVersion: 20,
-      pythonLayerVersion: 28,
-      addLayers: true,
-      forwarderArn: "forwarder-arn",
-      enableDatadogTracing: true,
-      flushMetricsToLogs: true,
-      site: "datadoghq.com",
-    });
-    const datadogCdk2 = new Datadog(stack, "Datadog2", {
-      nodeLayerVersion: 20,
-      pythonLayerVersion: 28,
-      addLayers: true,
-      forwarderArn: "forwarder-arn2",
-      enableDatadogTracing: true,
-      flushMetricsToLogs: true,
-      site: "datadoghq.com",
-    });
-    datadogCdk.addLambdaFunctions([nodeLambda]);
-    datadogCdk2.addLambdaFunctions([pythonLambda]);
-
-    const subscriptionFilters = findDatadogSubscriptionFilters(stack);
-    expect(nodeLambda.logGroup.node.tryFindChild(subscriptionFilters[0].id)).not.toBeUndefined();
-    expect(pythonLambda.logGroup.node.tryFindChild(subscriptionFilters[1].id)).not.toBeUndefined();
-    expect(subscriptionFilters[0].destinationArn).not.toEqual(subscriptionFilters[1].destinationArn);
-  });
-
-  it("Produces stable log subscription resource ids", () => {
-    const createStack = (functionId: string) => {
-      const app = new cdk.App();
-      const stack = new cdk.Stack(app, "stack", {
-        env: {
-          region: "us-gov-east-1",
-        },
-      });
-      const pythonLambda = new lambda.Function(stack, functionId, {
-        runtime: lambda.Runtime.PYTHON_3_7,
-        code: lambda.Code.fromAsset("test"),
-        handler: "hello.handler",
-      });
-      addForwarder(stack, [pythonLambda], "forwarder-arn");
-
-      return stack;
-    };
-
-    const initialStack = createStack("PythonHandler");
-    const identicalStack = createStack("PythonHandler");
-    const differentStack = createStack("OtherHandler");
-
-    const [initialStackSubscription] = findDatadogSubscriptionFilters(initialStack);
-    const [identicalStackSubscription] = findDatadogSubscriptionFilters(identicalStack);
-    const [differentStackSubscription] = findDatadogSubscriptionFilters(differentStack);
-
-    expect(initialStackSubscription.id).toEqual(identicalStackSubscription.id);
-    expect(initialStackSubscription.id).not.toEqual(differentStackSubscription.id);
-  });
 });
+
 describe("applyLayers", () => {
   it("if addLayers is not given, layer is added", () => {
     const app = new cdk.App();
